Add google login link to home page
Will move to sign in page eventually
