--- client/src/components/GoogleLogin.jsx
@@ -1,21 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import GoogleLogin from 'react-google-login';
-// or
-import { GoogleLogin } from 'react-google-login';
- 
- 
-const responseGoogle = (response) => {
-  console.log(response);
-}
- 
-ReactDOM.render(
-  <GoogleLogin
-    clientId="218365048814-n2ukcbljenpiqdr0n978dm663mkc4ofe.apps.googleusercontent.com"
-    buttonText="Login"
-    onSuccess={responseGoogle}
-    onFailure={responseGoogle}
-    cookiePolicy={'single_host_origin'}
-  />,
-  document.getElementById('googleButton')
-);
