--- client/src/components/GoogleLogout.jsx
@@ -1,7 +1,0 @@
-    import { GoogleLogout } from 'react-google-login';
-    <GoogleLogout
-      clientId="658977310896-knrl3gka66fldh83dao2rhgbblmd4un9.apps.googleusercontent.com"
-      buttonText="Logout"
-      onLogoutSuccess={logout}
-    >
-    </GoogleLogout>
