+++ client/src/views/Home.jsx
@@ -3,7 +3,11 @@ import {Link} from 'react-router-dom';
 import "../styles/index.scss";
 import Slideshow from "../components/Slideshow";
 import HomeGrid from "../components/HomeGrid/HomeGrid";
+import GoogleLogin from 'react-google-login';
 
+const responseGoogle = (response) => {
+                       console.log(response);
+                }
 
 class Home extends React.Component {
 
@@ -28,7 +32,16 @@ class Home extends React.Component {
 				{/* Tag line boxes  */}
 				<HomeGrid className="cardGrid"/>
 				<br/>
-
+                
+                
+    <GoogleLogin
+    clientId="218365048814-n2ukcbljenpiqdr0n978dm663mkc4ofe.apps.googleusercontent.com"
+    buttonText="Login"
+    onSuccess={responseGoogle}
+    onFailure={responseGoogle}
+    cookiePolicy={'single_host_origin'}
+  />,
+                
 				{/* Get involved box  */}
 				<div className="page-element">
 					<h1 className="title" id="get-involved-title">
@@ -53,7 +66,7 @@ class Home extends React.Component {
 					</p>
 				</div>
 
-				<div class="g-signin2" data-onsuccess="onSignIn"></div>
+
 
 			</div>
 		)
